Moved tests, minor refactors, getAllFlags, Updated docs and added npmignore
