+++ lib/index.js
@@ -12,13 +12,11 @@ class Draconarius {
    * @memberof Draconarius
    */
   constructor(flags) {
-    if (!flags || !(flags instanceof Object) || Array.isArray(flags)) {
+    if (!isObject(flags)) {
       throw new TypeError(`Draconarius constructor expects flags to be an object but received \`${JSON.stringify(flags)}\``);
     }
 
     this.flags = flags;
-
-    this.isEnabled = this.isEnabled.bind(this);
   }
 
   /**
@@ -34,7 +32,7 @@ class Draconarius {
   /**
    * _checkFlag - Checks if a flag is enabled
    * @private
-   * @param {string} flags Flag that need to check
+   * @param {string} flag Flag which need to checked
    * @param {array} args Optional arguments to function flags
    * @returns {boolean}
    * @memberof Draconarius
@@ -65,7 +63,7 @@ class Draconarius {
    *
    * _setFlag
    * @private
-   * @param {string} flag The flag or flags you want to check
+   * @param {string} flag
    * @param {function|boolean} value
    * @memberof Draconarius
    */
@@ -108,7 +106,7 @@ class Draconarius {
 
   /**
    *
-   * updateFlag
+   * updateFlag - Update flag(s)
    * @param {string|array} flag The flag or flags you want to check
    * @param {function|boolean} value
    * @memberof Draconarius
@@ -119,6 +117,16 @@ class Draconarius {
     }
 
     return this._setFlag(flag, value);
+  }
+
+  /**
+   *
+   * getAllFlags - Get all the flags currently set
+   * @returns {object} All flags
+   * @memberof Draconarius
+   */
+  getAllFlags() {
+    return this.flags;
   }
 }
